+++ src/App.jsx
@@ -20,15 +20,6 @@ import LoggedInRoute from "./firebase/LoggedInRoute";
 
 export default function App() {
 	const dispatch = useDispatch();
-	const user = useContext(AuthContext);
-	console.log(user)
-
-	const get_todayMidnight = () => {
-		const TODAY_MIDNIGHT = new Date();
-		TODAY_MIDNIGHT.setHours(0);
-		TODAY_MIDNIGHT.setMinutes(0);
-		return TODAY_MIDNIGHT.setSeconds(0);
-	};
 
 	useEffect(() => {
 		firebase
@@ -41,18 +32,6 @@ export default function App() {
 				});
 				dispatch(fetch_dot(RESPONSE));
 			});
-
-		// firebase
-		// 	.firestore()
-		// 	.collection("dots")
-		// 	.where("userId", "==", user.uid)
-		// 	.where("createdAt", ">=", new Date(get_todayMidnight()))
-		// 	.onSnapshot((snapshot) => {
-		// 		const todayDot = snapshot.docs.map((doc) => {
-		// 			return doc.data();
-		// 		});
-		// 		dispatch(fetch_today_dot(todayDot))
-		// 	});
 	}, []);
 
 	return (
